enhance: auctionTimerButton component seperate #14
