+++ src/Components/index.ts
@@ -6,3 +6,4 @@ export * from './Footer/Footer';
 export * from './ProductHeader/ProductHeader';
 export * from './ItemListCell/ItemListCell';
 export * from './ImageSlick/ImageSlick';
+export * from './AuctionTimerButton/AuctionTimerButton';
